Buscar_trabajo/requirements.txt
requests
gspread
google-auth
google-auth-oauthlib
gspread-formatting
